Added the Script.py file and updated the node.py class methods.
The script.py file is to create my own 'script' language since I am building a blockchain with python. I created a Stack class and a StackScript class. The Stack class is basic with two methods, where as the StackScript class inherits those methods from stack and adds 4 other methods for building and validating the scripting functions of the blockchain